refactor(integrations): registry-driven, category-grouped layout
Splits the monolithic IntegrationsApp into a registry of descriptors plus
per-card files, organised into category sections.

Categories
- External Agents: OpenClaw, Hermes (gateways that import existing agents)
- Channels: Telegram (inbound messaging streams as channels)

Why
Adding a new gateway used to require touching IntegrationsApp directly.
With a registry, a new integration is a 4-line descriptor + a card file —
the app re-renders without further edits and the category section appears
automatically. The category headers also give users a stable mental model
for where to look as more integrations land.

Files
- components/apps/integrations/types.ts: IntegrationCategory union,
  INTEGRATION_CATEGORIES metadata, IntegrationDescriptor + Context shapes.
- components/apps/integrations/CardShell.tsx: shared visual primitive.
- components/apps/integrations/{OpenClaw,Hermes,Telegram}Card.tsx: one
  card per integration, owning its own state and submit logic.
- components/apps/integrations/registry.tsx: the single source of truth
  for what renders in the Integrations app.
- components/apps/IntegrationsApp.tsx: trimmed to category-section
  rendering of the registry; no per-integration knowledge.

Tests
- registry.test.tsx: locks in id uniqueness, category membership, per-
  category coverage, and the build-gate (gateway_kinds=[]) suppression
  contract.
